requirements.txt
tqdm==4.61.0
camelot-py==0.8.2
tabula-py=2.2.0